HPCC-34920 Revert Add v9 activity tests
This reverts commit d5f4f7bc615d71db4db4e07008773c4061e3f2a4.
